QuantumStateTransfer

Documentation for QuantumStateTransfer.

QuantumStateTransfer.OptimizedStateTransfer

QuantumStateTransfer.QubitPairTransfer

QuantumStateTransfer.UnitaryEvolution

QuantumStateTransfer.optimized_state_transfer

QuantumStateTransfer.qubit_pair_transfer

QuantumStateTransfer.track_qubit_amplitude

QuantumStateTransfer.unitary_evolution

QuantumStateTransfer.OptimizedStateTransfer — Type

OptimizedStateTransfer

A struct to store data on optimized state transfer between all qubit pairs in a network.

Fields

adjacency_matrix::AbstractMatrix{<:Real}: The network in adjacency_matrix format.

exhibits_pst::Bool: Whether the quantum network exhibits perfect state transfer

between any pair of qubits.

qubit_pairs::Base.Generator: A lazily evaluated generator of structs containing

data on state transfer between every pair of qubits in the network.

QuantumStateTransfer.QubitPairTransfer — Type

QubitPairTransfer

A struct to store data on optimized state transfer between two qubits in a quantum network.

Fields

source::Int: The index of the source qubit.

dest::Int: The index of the target qubit.

is_pst::Bool: Whether there is perfect state transfer from the source qubit to the

target qubit (i.e., whether maximum_fidelity is sufficiently close to 1).

maximum_fidelity::Float64: The (numerically approximated) maximum transfer fidelity from

the source qubit to the target qubit.

optimal_time::Float64: The time at which the maximum transfer fidelity is achieved.

QuantumStateTransfer.UnitaryEvolution — Type

UnitaryEvolution

A struct to store data on the unitary evolution of a quantum network.

Fields

adjacency_matrix::AbstractMatrix{<:Real}: The network in adjacency_matrix format.

time_steps::AbstractVector{<:Real}: The times at which to track state amplitudes.

transfer_amplitudes::Array{ComplexF64, 3}: The amplitudes of state transfer between

qubits. The first dimension corresponds to the time step, the second corresponds to the target qubit, and the third corresponds to the source qubit.

transfer_fidelities::Array{Float64, 3}: The magnitudes of the transfer amplitudes

squared, representing wave function probabilities rather than amplitudes.

QuantumStateTransfer.optimized_state_transfer — Method

optimized_state_transfer(adj_mat::AbstractMatrix{<:Real}; kwargs...)
optimized_state_transfer(graph::AbstractGraph{Int}; kwargs...)

Add later

Arguments

Add later

Keywords

See qubit_pair_transfer documentation for more information. (The set of valid keyword arguments is exactly the same for both functions, as qubit_pair_transfer is a helper function for optimized_state_transfer.)

Returns

Add later

Examples

Add later ```

QuantumStateTransfer.qubit_pair_transfer — Method

qubit_pair_transfer(
    adj_mat::AbstractMatrix{<:Real}, source::Int, dest::Int;
    min_time::Real=0,
    max_time::Real=12.566370614359172,
    tol::Real=0.01,
)
qubit_pair_transfer(graph::AbstractGraph{Int}, source::Int, dest::Int; kwargs...)

Find the maximized fidelity of state transfer between two qubits and the associated time.

Arguments

adj_mat::AbstractMatrix{<:Real}: The quantum network in adjacency_matrix format.

graph::AbstractGraph{Int}: The network in graph format. (Only passed when adj_mat is

not provided.)

source::Int: The index of the source qubit (assumed to have an initial state of the

source-th standard basis vector).

dest::Int: The index of the target qubit (assumed to have an initial state of the

dest-th standard basis vector).

Keywords

min_time::Real=0: The minimum time at which state transfer is considered.

max_time::Real=12.566370614359172: The maximum time at which state transfer is considered.

tol::Real=0.01: The margin of error for determining whether perfect state transfer

occurs between the source and target.

Returns

QubitPairTransfer: A struct containing data on the optimized state transfer. Also

indicates whether PST was found between the two qubits.

Examples

```jldoctest julia> using Graphs: Graph

julia> C4_adj = BitMatrix([0 1 0 1; # Cycle graph on 4 vertices (adj. matrix format) 1 0 1 0; 0 1 0 1; 1 0 1 0]);

julia> C4graph = Graph(C4adj); # Cycle graph on 4 vertices (graph format)

julia> qubitpairtransfer(C4_adj, 1, 2) # There is no PST from node 1 to node 2 QubitPairTransfer(1, 2, false, 0.25, 2.35619)

julia> qubitpairtransfer(C4_graph, 1, 3) # There is PST from node 1 to node 3 over time π/2 QubitPairTransfer(1, 3, true, 1.0, 1.5708)

QuantumStateTransfer.track_qubit_amplitude — Method

track_qubit_amplitude(
    adj_mat::AbstractMatrix{<:Real}, source::Int;
    dests::AbstractVector{Int}=1:size(adj_mat, 1),
    time_steps::AbstractVector{<:Real}=0:(π / 200):5,
)
track_qubit_amplitude(graph::AbstractGraph{Int}, source::Int; kwargs...)

Tracks the wave function amplitude of a qubit in a quantum network over time.

Arguments

adj_mat::AbstractMatrix{<:Real}: The quantum network in adjacency_matrix format.

graph::AbstractGraph{Int}: The network in graph format. (Only passed when adj_mat is

not provided.)

source::Int: The qubit whose wave function amplitude to track.

Keywords

dests::AbstractVector{Int}=1:size(adj_mat, 1): The qubits at which to track the wave

function amplitude of source.

time_steps::AbstractVector{<:Real}=0:(π / 200):5: The times at which to track the wave

function amplitude of source.

Returns

Matrix{ComplexF64}: The transfer amplitudes of source to each qubit in dests at each

time step. The first dimension corresponds to the time step, and the second corresponds to the target qubit.

Examples

julia> using Graphs: wheel_graph

julia> g = wheel_graph(8); source = 3; dests = [2, 4, 7]; time_steps = 0:1:5;

julia> amps_3_to_247 = track_qubit_amplitude(g, source, dests=dests, time_steps=time_steps)
6×3 Matrix{ComplexF64}:
        0.0-0.0im               0.0-0.0im               0.0-0.0im
 -0.0282559+0.340869im   -0.0282559+0.340869im   0.00691979-0.364799im
  0.0270747+0.159635im    0.0270747+0.159635im     0.353068-0.20453im
  -0.248756-0.288904im    -0.248756-0.288904im     0.291226-0.129163im
  -0.158849+0.0938213im   -0.158849+0.0938213im   -0.210858+0.121758im
   0.500422+0.160662im     0.500422+0.160662im    -0.259143-0.0882654im

QuantumStateTransfer.unitary_evolution — Function

unitary_evolution(
    adj_mat::AbstractMatrix{<:Real}, time_steps::AbstractVector{<:Real}=0:(π / 200):5,
)
unitary_evolution(
    graph::AbstractGraph{Int}, time_steps::AbstractVector{<:Real}=0:(π / 200):5,
)

Simulates unitary evolution on a quantum network in the idealized absence of noise.

Arguments

adj_mat::AbstractMatrix{<:Real}: The quantum network in adjacency_matrix format.

graph::AbstractGraph{Int}: The network in graph format. (Only passed when adj_mat is

not provided.)

Keywords

time_steps::AbstractVector{<:Real}=0:(π / 200):5: The times at which to track state

transfer between qubits.

Returns

UnitaryEvolution: A struct storing data on the unitary evolution of the quantum network.

Most notably, contains the amplitude and fidelity of state transfer between each pair of qubits at every time step.

Examples

julia> using Graphs: complete_bipartite_graph

julia> g = complete_bipartite_graph(1, 2); time_steps = LinRange(0, π, 7);

julia> evolution = unitary_evolution(g, time_steps); # Simulate state transfer on K₁₂

julia> evolution.adjacency_matrix # Adjacency matrix of the complete bipartite graph K₁₂
3×3 BitMatrix:
 0  1  1
 1  0  0
 1  0  0

julia> evolution.time_steps # Time steps at which to measure state transfer
7-element LinRange{Float64, Int64}:
 0.0, 0.523599, 1.0472, 1.5708, 2.0944, 2.61799, 3.14159

julia> evolution.transfer_amplitudes # State transfer amplitudes between qubits
7×3×3 Array{ComplexF64, 3}:
[:, :, 1] =
       1.0-0.0im  0.0-0.0im       0.0-0.0im
  0.738144-0.0im  0.0+0.477044im  0.0+0.477044im
 0.0897146-0.0im  0.0+0.704255im  0.0+0.704255im
   -0.6057-0.0im  0.0+0.56264im   0.0+0.56264im
 -0.983903-0.0im  0.0+0.126364im  0.0+0.126364im
 -0.846825-0.0im  0.0-0.37609im   0.0-0.37609im
 -0.266255-0.0im  0.0-0.681582im  0.0-0.681582im

[:, :, 2] =
 0.0-0.0im             1.0-0.0im        0.0-0.0im
 0.0+0.477044im   0.869072-0.0im  -0.130928-0.0im
 0.0+0.704255im   0.544857-0.0im  -0.455143-0.0im
 0.0+0.56264im     0.19715-0.0im   -0.80285-0.0im
 0.0+0.126364im  0.0080487-0.0im  -0.991951-0.0im
 0.0-0.37609im   0.0765877-0.0im  -0.923412-0.0im
 0.0-0.681582im   0.366872-0.0im  -0.633128-0.0im

[:, :, 3] =
 0.0-0.0im             0.0-0.0im        1.0-0.0im
 0.0+0.477044im  -0.130928-0.0im   0.869072-0.0im
 0.0+0.704255im  -0.455143-0.0im   0.544857-0.0im
 0.0+0.56264im    -0.80285-0.0im    0.19715-0.0im
 0.0+0.126364im  -0.991951-0.0im  0.0080487-0.0im
 0.0-0.37609im   -0.923412-0.0im  0.0765877-0.0im
 0.0-0.681582im  -0.633128-0.0im   0.366872-0.0im

julia> evolution.transfer_fidelities # State transfer fidelities between qubits
7×3×3 Array{Float64, 3}:
[:, :, 1] =
 1.0        0.0        0.0
 0.544857   0.227571   0.227571
 0.0080487  0.495976   0.495976
 0.366872   0.316564   0.316564
 0.968064   0.0159678  0.0159678
 0.717112   0.141444   0.141444
 0.0708919  0.464554   0.464554

[:, :, 2] =
 0.0        1.0         0.0
 0.227571   0.755287    0.0171421
 0.495976   0.296869    0.207155
 0.316564   0.0388681   0.644568
 0.0159678  6.47816e-5  0.983967
 0.141444   0.00586567  0.85269
 0.464554   0.134595    0.400851

[:, :, 3] =
 0.0        0.0        1.0
 0.227571   0.0171421  0.755287
 0.495976   0.207155   0.296869
 0.316564   0.644568   0.0388681
 0.0159678  0.983967   6.47816e-5
 0.141444   0.85269    0.00586567
 0.464554   0.400851   0.134595